requirements.txt
torch>=1.9.0
torch-geometric>=2.0.0
numpy>=1.19.0
pandas>=1.3.0
scikit-learn>=0.24.0
biopython>=1.79
matplotlib>=3.4.0
seaborn>=0.11.0
tqdm>=4.62.0